+++ adminutils/widgets.py
@@ -1,6 +1,8 @@
 import json
 
 from django.contrib.admin.templatetags.admin_list import _boolean_icon
+from django.contrib.admin.widgets import AdminDateWidget
+from django.contrib.postgres.forms import RangeWidget
 from django.core.serializers.json import DjangoJSONEncoder
 from django.utils import html
 from django.utils.safestring import mark_safe
@@ -57,3 +59,10 @@ def admin_detail_link(instance, text=None, bold=False, action="change"):
     text = str(instance) if text is None else text
     style = "font-weight: bold;" if bold else ""
     return html.format_html('<a href="{}" style="{}">{}</a>', url, style, text)
+
+
+class AdminDateRangeWidget(RangeWidget):
+    template_name = "admin/rangefield.html"
+
+    def __init__(self, attrs=None):
+        super().__init__(AdminDateWidget, attrs=attrs)